svn/branches : Added support for USB-Devices --------------------------------------------------------- For testing uncomment #define USB_DEVICE in defines.h
Requirements: USB-HDD with 2 primary partitions
( e.g. 4GB FAT32 and 196GB WBFS )
- copy content of SDcard into root of FAT-partition
- don't insert SDcard, it's preferred...

modified:
- CoverFlow.pnproj
- cheats.h
- core/usbstorage.h
- coverflow.h
- defines.h
- core/fst.c
- core/subsystem.c
- core/usbstorage.c
- gameinfo.c
- main.c
- storage.c

added:
- core/fatmounter.h
- core/fatmounter.c

+++ source/libwbfs/libwbfs.c
@@ -542,7 +542,7 @@ error:
 }
 u32 wbfs_extract_file(wbfs_disc_t*d, char *path);
 
-u32 wbfs_estimate_disc(
+u64 wbfs_estimate_disc(
 		wbfs_t *p, read_wiidisc_callback_t read_src_wii_disc,
 		void *callback_data,
 		partition_selector_t sel)
@@ -598,5 +598,5 @@ error:
 	if (info)
 		wbfs_iofree(info);
 
-	return tot * ((p->wbfs_sec_sz / p->hd_sec_sz) * 512);
+	return (u64) tot * ((p->wbfs_sec_sz / p->hd_sec_sz) * 512);
 }
